+++ utils/get_functions.py
@@ -21,7 +21,19 @@ logging.basicConfig(
 )
 logger = logging.getLogger(__name__)
 
-SET_FAV_MULTI, CHOOSE_MODE, GET_LINK, IS_COVER_QUICK, GET_DST, COOK_ID, REGEX_IN, REGEX_GET_DST = range(8)
+(
+    SET_FAV_MULTI,
+    CHOOSE_MODE,
+    GET_LINK,
+    IS_COVER_QUICK,
+    GET_DST,
+    COOK_ID,
+    REGEX_IN,
+    REGEX_GET_DST,
+    COOK_FAV_TO_SIZE,
+    COOK_FAV_PURGE,
+    COOK_ID_DEDU,
+) = range(11)
 
 regex1 = r"[-\w]{11,}"
 regex2 = r"[-\w]"