+++ utils/get_set.py
@@ -13,7 +13,19 @@ from telegram import ParseMode
 from utils.load import _lang, _text
 from drive.gdrive import GoogleDrive as _gd
 
-SET_FAV_MULTI, CHOOSE_MODE, GET_LINK, IS_COVER_QUICK, GET_DST, COOK_ID, REGEX_IN, REGEX_GET_DST = range(8)
+(
+    SET_FAV_MULTI,
+    CHOOSE_MODE,
+    GET_LINK,
+    IS_COVER_QUICK,
+    GET_DST,
+    COOK_ID,
+    REGEX_IN,
+    REGEX_GET_DST,
+    COOK_FAV_TO_SIZE,
+    COOK_FAV_PURGE,
+    COOK_ID_DEDU,
+) = range(11)
 
 pick_quick = []
 pick_fav = []