+++ utils/task_box.py
@@ -5,6 +5,7 @@ import time, pymongo, re
 from utils import load,task_payload as _payload
 from telegram.ext import ConversationHandler
 from utils.load import _lang, _text
+from telegram import ParseMode
 
 future = load.db_counters.find_one({"_id": "task_list_id"})
 future_id = 0
@@ -24,6 +25,16 @@ def cook_task_to_db(update, context, tmp_task_list):
         item["error"] = 0
         item["create_time"] = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())
         item["finished_time"] = ""
+        item["start_time"] = ""
+        item["task_current_prog_num"] = 0
+        item["task_total_prog_num"] = 0
+        item["task_current_prog_size"] = 0
+        item["task_total_prog_size"] = 0
+        item["task_current_prog_size_tail"] = ""
+        item["task_total_prog_size_tail"] = ""
+        item["dst_endpoint_link"] = ""
+        item["dst_endpoint_id"] = ""
+        item["is_reset"] = 0
 
     insert_callback = load.task_list.insert_many(tmp_task_list)
     if insert_callback.inserted_ids:
@@ -35,6 +46,7 @@ def cook_task_to_db(update, context, tmp_task_list):
 
 def taskinfo(update, context):
     entry_cmd = update.effective_message.text
+    match_cmd = re.search(r'^\/task ([1-9]\d*)$', entry_cmd, flags=re.I)
     if " " in entry_cmd:
         entry_cmd = entry_cmd.replace(" ","")
 
@@ -90,6 +102,61 @@ def taskinfo(update, context):
             )
 
             return ConversationHandler.END
+
+    elif match_cmd:
+        limit_query = load.db_counters.find_one({"_id":"task_list_id"})
+        check_query = match_cmd.group(1)
+
+        if int(check_query) <= limit_query['future_id']:
+
+            check_task = load.task_list.find_one({"_id": int(check_query)})
+
+            if check_task["status"] == 1:
+                if "task_current_prog_num" and "task_current_prog_size_tail" in check_task:
+                    if check_task["error"] == 0:
+                        task_status_now = _text[_lang]["done"]
+                    elif check_task["error"] == 1:
+                        task_status_now == _text[_lang]["is_interrupted_error"]
+                    elif check_task["error"] == 9:
+                        task_status_now == _text[_lang]["is_killed_by_user"]
+
+                    check_msg = (
+                        _text[_lang]["current_task_id"]
+                        + str(check_query)
+                        + "\n"
+                        + '<a href="{}">{}</a>'.format(check_task["dst_endpoint_link"], check_task["src_name"])
+                        + "\n"
+                        + _text[_lang]["task_status"]
+                        + task_status_now
+                        + "\n"
+                        + _text[_lang]["total_file_num"]
+                        + str(check_task["task_current_prog_num"])
+                        + "/"
+                        + str(check_task["task_total_prog_num"])
+                        + "\n"
+                        + _text[_lang]["total_file_size"]
+                        + str(check_task["task_current_prog_size"])
+                        + check_task["task_current_prog_size_tail"]
+                        + "/"
+                        + str(check_task["task_total_prog_size"])
+                        + check_task["task_total_prog_size_tail"]
+                    )
+
+                else:
+                    check_msg = _text[_lang]["support_error"]
+
+            elif check_task["status"] == 0:
+                check_msg = _text[_lang]["task_is_in_queue"] + "\n" + _text[_lang]["finished_could_be_check"]
+
+            elif check_task["status"] == 2:
+                check_msg = _text[_lang]["doing"] + "\n" + _text[_lang]["finished_could_be_check"]
+
+            update.effective_message.reply_text(check_msg,parse_mode=ParseMode.HTML,disable_web_page_preview=True)
+
+        else:
+            update.effective_message.reply_text(_text[_lang]["over_limit_to_check"])
+
+
     else:
         return ConversationHandler.END
 
@@ -99,7 +166,7 @@ def task_reset(update, context):
 
     if "/reset" == entry_cmd:
         check_query = load.db_counters.find_one({"_id": "last_task"})
-        load.task_list.update_one({"_id": check_query['task_id']}, {"$set": {"status": 0,}})
+        load.task_list.update_one({"_id": check_query['task_id']}, {"$set": {"status": 0,"is_reset": 1}})
         update.effective_message.reply_text(
             _text[_lang]["reset_successful"].replace("replace",str(check_query['task_id']))
         )
@@ -110,7 +177,7 @@ def task_reset(update, context):
 
         if int(check_query) <= limit_query['future_id']:
 
-            load.task_list.update_one({"_id": int(check_query)}, {"$set": {"status": 0}})
+            load.task_list.update_one({"_id": int(check_query)}, {"$set": {"status": 0,"is_reset": 1}})
             update.effective_message.reply_text(
                 _text[_lang]["reset_successful"].replace("replace",check_query)
             )