+++ utils/task_payload.py
@@ -7,8 +7,10 @@ from multiprocessing import Process as _mp, Manager
 from utils.load import _lang, _text
 import subprocess
 from threading import Timer
-#from telegram import Bot
-#from telegram.utils.request import Request as TGRequest
+from drive.gdrive import GoogleDrive as _gd
+from telegram import ParseMode
+from telegram.utils.request import Request as TGRequest
+from telegram import Bot
 
 myclient = pymongo.MongoClient(
     f"{load.cfg['database']['db_connect_method']}://{load.user}:{load.passwd}@{load.cfg['database']['db_addr']}",
@@ -17,8 +19,13 @@ myclient = pymongo.MongoClient(
 )
 mydb = myclient[load.cfg["database"]["db_name"]]
 task_list = mydb["task_list"]
-db_counters = mydb['counters']
+db_counters = mydb["counters"]
+
 _cfg = load.cfg
+
+request = TGRequest(con_pool_size=8)
+bot = Bot(token=f"{_cfg['tg']['token']}", request=request)
+
 message_info = ""
 prog_bar = ""
 current_working_file = ""
@@ -27,16 +34,20 @@ now_elapsed_time = ""
 context_old = ""
 icopyprocess = subprocess.Popen
 interruption = 0
+dst_id = ""
+src_name = ""
 
 
 def task_buffer(ns):
+    global dst_id
+    global src_name
     while True:
         wait_list = task_list.find({"status": 0})
         for task in wait_list:
             if _cfg["general"]["cloner"] == "fclone":
-                flags = ['--drive-server-side-across-configs', '--check-first']
+                flags = ["--drive-server-side-across-configs", "--check-first"]
             else:
-                flags = ['--drive-server-side-across-configs']
+                flags = ["--drive-server-side-across-configs"]
             command = []
 
             cloner = _cfg["general"]["cloner"]
@@ -85,10 +96,7 @@ def task_buffer(ns):
 def task_process(chat_id, command, task, ns):
     # mark is in processing in db
     task_list.update_one({"_id": task["_id"]}, {"$set": {"status": 2,}})
-    db_counters.update({"_id": "last_task"},{"task_id": task["_id"]},upsert=True)
-    #request = TGRequest(con_pool_size=8)
-    #bot = Bot(token=f"{_cfg['tg']['token']}", request=request)
-    bot = load.bot
+    db_counters.update({"_id": "last_task"}, {"task_id": task["_id"]}, upsert=True)
     chat_id = chat_id
     message = bot.send_message(chat_id=chat_id, text=_text[_lang]["ready_to_task"])
     message_id = message.message_id
@@ -119,7 +127,7 @@ def task_process(chat_id, command, task, ns):
             r"Transferred:\s+(\d+) / (\d+), (\d+)%(?:,\s*([\d.]+\sFiles/s))?"
         )
         regex_total_size = (
-            r"Transferred:[\s]+([\d.]+\s*[kMGTP]?) / ([\d.]+[\s]?[kMGTP]?Bytes),"
+            r"Transferred:[\s]+([\d.]+\s*)([kMGTP]?) / ([\d.]+[\s]?)([kMGTP]?Bytes),"
             r"\s*(?:\-|(\d+)\%),\s*([\d.]+\s*[kMGTP]?Bytes/s),\s*ETA\s*([\-0-9hmsdwy]+)"
         )
 
@@ -138,10 +146,12 @@ def task_process(chat_id, command, task, ns):
                 task_in_file_speed = task_total_files.group(4)
 
             if task_total_size:
-                task_current_prog_size = task_total_size.group(1)
-                task_total_prog_size = task_total_size.group(2)
-                task_in_size_speed = task_total_size.group(4)
-                task_eta_in_file = task_total_size.group(5)
+                task_current_prog_size = task_total_size.group(1).strip()
+                task_current_prog_size_tail = task_total_size.group(2)
+                task_total_prog_size = task_total_size.group(3).strip()
+                task_total_prog_size_tail = task_total_size.group(4)
+                task_in_size_speed = task_total_size.group(6)
+                task_eta_in_file = task_total_size.group(7)
 
             if task_elapsed_time:
                 global now_elapsed_time
@@ -184,8 +194,10 @@ def task_process(chat_id, command, task, ns):
             + "\n\n"
             + _text[_lang]["task_files_size"]
             + str(task_current_prog_size)
+            + task_current_prog_size_tail
             + "/"
             + str(task_total_prog_size)
+            + task_total_prog_size_tail
             + "\n"
             + _text[_lang]["task_files_num"]
             + str(task_current_prog_num)
@@ -246,6 +258,13 @@ def task_process(chat_id, command, task, ns):
 
     old_working_file = ""
     finished_time = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())
+
+    dst_endpoint_id = _gd.get_dst_endpoint_id(_gd(), dst_id, src_name)
+    if dst_endpoint_id:
+        dst_endpoint_link = r"https://drive.google.com/open?id={}".format(
+            dst_endpoint_id["id"]
+        )
+
     if ns.x == 0:
         time.sleep(5)
         prog_bar = _bar.status(100)
@@ -259,81 +278,89 @@ def task_process(chat_id, command, task, ns):
             + _text[_lang]["current_task_id"]
             + str(task["_id"])
             + "\n\n"
-            + message_info
+            + message_info.replace(
+                "    ┕─📃" + task["src_name"],
+                "    ┕─📃"
+                + '<a href="{}">{}</a>'.format(dst_endpoint_link, task["src_name"]),
+            )
             + "\n"
             + _text[_lang]["task_finished_time"]
             + finished_time
             + "\n"
             + _text[_lang]["elapsed_time"]
             + str(now_elapsed_time),
+            parse_mode=ParseMode.HTML,
+            disable_web_page_preview=True,
         )
-        task_list.update_one(
-            {"_id": task["_id"]},
-            {
-                "$set": {
-                    "status": 1,
-                    "start_time": start_time,
-                    "finished_time": finished_time,
-                    "task_current_prog_num": task_current_prog_num,
-                    "task_total_prog_num": task_total_prog_num,
-                    "task_current_prog_size": task_current_prog_size,
-                    "task_total_prog_size" : task_total_prog_size,
-                }
-            },
-        )
+        check_is_reset = task_list.find_one({"_id": task["_id"]})
+        if check_is_reset['is_reset'] == 0:
+            task_list.update_one(
+                {"_id": task["_id"]},
+                {
+                    "$set": {
+                        "status": 1,
+                        "start_time": start_time,
+                        "finished_time": finished_time,
+                        "task_current_prog_num": int(task_current_prog_num),
+                        "task_total_prog_num": int(task_total_prog_num),
+                        "task_current_prog_size": float(task_current_prog_size),
+                        "task_total_prog_size": float(task_total_prog_size),
+                        "task_current_prog_size_tail": task_current_prog_size_tail,
+                        "task_total_prog_size_tail": task_total_prog_size_tail,
+                        "dst_endpoint_link": dst_endpoint_link,
+                        "dst_endpoint_id": dst_endpoint_id["id"],
+                    }
+                },
+            )
 
-    interrupted_msg = (
-        _text[_lang]["task_src_info"]
-        + "\n"
-        + "📃"
-        + task["src_name"]
-        + "\n"
-        + "----------------------------------------"
-        + "\n"
-        + _text[_lang]["task_dst_info"]
-        + "\n"
-        + "📁"
-        + task["dst_name"]
-        + ":"
-        + "\n"
-        + "    ┕─📃"
-        + task["src_name"]
-        + "\n"
-        + "----------------------------------------"
-        + "\n\n"
-        + _text[_lang]["task_files_size"]
-        + str(task_current_prog_size)
-        + "/"
-        + str(task_total_prog_size)
-        + "\n"
-        + _text[_lang]["task_files_num"]
-        + str(task_current_prog_num)
-        + "/"
-        + str(task_total_prog_num)
-        + "\n\n"
-        + str(task_percent)
-        + "%"
-        + str(prog_bar)
-    )
+        if check_is_reset['is_reset'] == 1:
+            task_list.update_one(
+                {"_id": task["_id"]},
+                {
+                    "$set": {
+                        "status": 1,
+                        "start_time": start_time,
+                        "finished_time": finished_time,
+                        "task_current_prog_num": int(task_current_prog_num) + check_is_reset['task_current_prog_num'],
+                        "task_total_prog_num": int(task_total_prog_num) + check_is_reset['task_total_prog_num'],
+                        "task_current_prog_size": 0,
+                        "task_total_prog_size": 0,
+                        "task_current_prog_size_tail": "",
+                        "task_total_prog_size_tail": "",
+                        "dst_endpoint_link": dst_endpoint_link,
+                        "dst_endpoint_id": dst_endpoint_id["id"],
+                    }
+                },
+            )
 
     if ns.x == 1:
         bot.edit_message_text(
             chat_id=chat_id,
             message_id=message_id,
             text=" ༺ ✪iCopy✪ ༻ \n"
-            +_text[_lang]["killed"]
+            + _text[_lang]["killed"]
             + " | "
             + "🏳️"
             + _text[_lang]["current_task_id"]
             + str(task["_id"])
             + "\n\n"
-            + interrupted_msg
+            + message_info.replace(
+                "\n\n" + _text[_lang]["task_start_time"] + start_time, ""
+            ).replace(
+                "\n"
+                + _text[_lang]["task_status"]
+                + "\n\n"
+                + str(task_in_size_speed)
+                + "  |  "
+                + str(task_in_file_speed),
+                "",
+            )
             + "\n"
             + _text[_lang]["is_killed_by_user"],
         )
 
         task_list.update_one(
-            {"_id": task["_id"]}, 
+            {"_id": task["_id"]},
             {
                 "$set": {
                     "status": 1,
@@ -341,9 +368,9 @@ def task_process(chat_id, command, task, ns):
                     "start_time": start_time,
                     "finished_time": finished_time,
                 }
-            }
+            },
         )
-    
+
     if interruption == 1:
         bot.edit_message_text(
             chat_id=chat_id,
@@ -355,13 +382,23 @@ def task_process(chat_id, command, task, ns):
             + _text[_lang]["current_task_id"]
             + str(task["_id"])
             + "\n\n"
-            + interrupted_msg
+            + message_info.replace(
+                "\n\n" + _text[_lang]["task_start_time"] + start_time, ""
+            ).replace(
+                "\n"
+                + _text[_lang]["task_status"]
+                + "\n\n"
+                + str(task_in_size_speed)
+                + "  |  "
+                + str(task_in_file_speed),
+                "",
+            )
             + "\n"
             + _text[_lang]["is_interrupted_error"],
         )
 
         task_list.update_one(
-            {"_id": task["_id"]}, 
+            {"_id": task["_id"]},
             {
                 "$set": {
                     "status": 1,
@@ -369,9 +406,8 @@ def task_process(chat_id, command, task, ns):
                     "start_time": start_time,
                     "finished_time": finished_time,
                 }
-            }
+            },
         )
-
 
     prog_bar = _bar.status(0)