+++ workflow/regex_workflow.py
@@ -4,7 +4,19 @@ from telegram.ext import CallbackQueryHandler, ConversationHandler
 from threading import Thread
 from multiprocessing import Process as _mp
 
-SET_FAV_MULTI, CHOOSE_MODE, GET_LINK, IS_COVER_QUICK, GET_DST, COOK_ID, REGEX_IN, REGEX_GET_DST = range(8)
+(
+    SET_FAV_MULTI,
+    CHOOSE_MODE,
+    GET_LINK,
+    IS_COVER_QUICK,
+    GET_DST,
+    COOK_ID,
+    REGEX_IN,
+    REGEX_GET_DST,
+    COOK_FAV_TO_SIZE,
+    COOK_FAV_PURGE,
+    COOK_ID_DEDU,
+) = range(11)
 
 src_name_list = []
 src_id_list = []