+++ docs/CHANGELOG.md
@@ -1,5 +1,56 @@
 # iCopy v0.2 CHANGELOG
 
+## version 0.2.0-beta.6
+
++ Update :  
+  + ADD : insert more details into Database and more initializated data  
+  + ADD : feedback dst endpoint link when task end normally  
+  + ADD : "/task id" only support the task which is start after v0.2.0b6  
+  + ADD : mark tasks that have been reset in the database  
+  + ADD : "/size id" & "/size fav"  
+  + ADD : "/purge" to empty shared drive trash bin  
+  + ADD : "/dedupe id" to dedupe task  
+  + ADD : Record the last time of size and dedupe  
+
++ Fix :
+  + FIX : Update RegEX Rules  
+
++ Root Command：
+
+  + start - nothing just say hello  
+  + menu - main entry point  
+quick - quick mode  
+copy - full mode  
+set - customize settings  
+task - task query  
+reset - restore task  
+size - just size task  
+dedupe - dedupe specified task  
+purge - delete files and folder in specified fav trash bin  
+cancel - cancel TG conversation  
+kill - kill task which is in processing  
+ver - check iCopy version  
+restart - restart iCopy  
+
+  + Child Command:
+
+    + set - customize settings  
+    ┖ set - batch way  
+    ┖ set rule - rules  
+    ┖ fav|quick +/- id - single way  
+    ┖ set purge - purge favorites  
+    + size - size query
+    ┖ size - size the shared resource  
+    ┖ size id - size specified task  
+    ┖ size fav - size specified favorites  
+    + task - task query  
+    ┖ task - task in processing  
+    ┖ task list - future 10 tasks  
+    ┖ task id - show the specified task
+    + reset - restore task  
+    ┖ reset - restore current task  
+    ┖ reset id  - restore the specified task  
+
 ## version 0.2.0-beta.5.1
 
 + Update :